## Before this commit:
We store the complete list of modules for each version. Therefore, for each new version, we add ~900 lines to `OUR_MODULES.json`, which is less maintainable and more difficult to review.

## After this commit:
We store the added and removed modules between the versions. This way, only a few lines are added for each update.

<details><summary>As an example, here is how 16.3 is defined.</summary>
<p>

```JSON
{
    "saas~16.3": {
        "+": [
            "account_peppol",
            "account_qr_code_emv",
            "l10n_cl_edi_website_sale",
            "l10n_do_reports",
            "l10n_dz_reports",
            "l10n_es_edi_facturae",
            "l10n_fr_hr_holidays",
            "l10n_fr_hr_work_entry_holidays",
            "l10n_ma_reports",
            "mrp_accountant",
            "pos_restaurant_preparation_display",
            "pos_self_order",
            "project_account",
            "project_todo",
            "sale_subscription_stock",
            "spreadsheet_dashboard_sale_subscription",
            "spreadsheet_dashboard_website_sale",
            "test_spreadsheet"
        ],
        "-": [
            "l10n_hk_hr_payroll_hsbc_autopay",
            "l10n_mx_edi_40",
            "l10n_mx_edi_extended_40",
            "l10n_mx_edi_stock_40",
            "l10n_mx_edi_stock_extended_40",
            "l10n_pe_edi_stock_20",
            "note",
            "pos_epson_printer_restaurant",
            "pos_restaurant_iot",
            "web_kanban_gauge",
            "website_event_crm_questions",
            "website_event_questions"
        ]
    }
}
``` 

</p>
</details>

## Pros
- It's easier to review when adding a new version, as we only see the added & removed modules.
- You can easily get the lifetime of a module with a simple <kbd>Ctrl + f</kbd> I.E. `knowledge` was added in `6.0`, removed in `9.0`, and added back in `16.0`.
- The file is smaller, and is not redundant

## Cons
- We need to compute the module list for a particular version, so we can't simply grab it from the file.
- We need to add the same logic in remote code, where the file was downloaded and used as-is.
- The more versions are added, the slower it will get

## Perf
<details><summary>Merging the diffs together can be done using sets operations, which is blazing fast in python.
</summary>
<p>

```python
def parse_version(version: str) -> float:
    return float(version.replace("saas~", ""))

def get_modules(target_version: str) -> set[str]:
    folder = Path(__file__).parent
    with open(folder / "OUR_MODULES_DIFF.json", "r") as f:
        modules_by_version = json.load(f)

    modules: set[str] = set()
    for version, mods in sorted(modules_by_version.items(), key=lambda v: parse_version(v[0])):
        modules = (modules - set(mods["-"])) | set(mods["+"])
        if version == target_version:
            break

    return modules
```

</p>
</details> 

The overhead is negligible, as it takes about the same amount of time.
|| Before | After |
|--------|--------|--------|
| **Best case** | 2.13 ms | 0.89 ms |
| **Worst case** | 2.13 ms | 2.99 ms |

## File size
The file is reduced from 369.65 KB to 69.74 KB (18.87%), or, 14025 lines to 2252 lines (16.06%).

## Extra
<details><summary>If you are desperately in need of a bash command to get the module list of a particular version, you can use this small script with JQ</summary>
<p>

[JQ](https://github.com/jqlang/jq) is a powerful json-parsing command line utility. 

Save the following as `get_modules.jq`:
```jq
def get_versions: .
    | keys
    | sort_by(. | sub("saas~"; "") | tonumber);


# Input $version using: `jq -f get_modules.jq file.json --arg version <VERSION>`
. as $gbl
    | get_versions
    | .[:index($version) + 1] as $versions
    | reduce .[] as $v ([]; (. - $gbl[$v]["-"]) + $gbl[$v]["+"])
    | sort
```

And run the command:
```shell
$ jq -f get_modules.jq ./OUR_MODULES.json --arg version "saas~16.3"
```

</p>
</details>